add json4vhdl.py
